optimized code with router.wrap implement server's gzip\session\cors in features update test code
