+++ router_test.go
@@ -1,8 +1,10 @@
 package dotweb
 
 import (
-	"testing"
+	"github.com/devfeel/dotweb/session"
 	"github.com/devfeel/dotweb/test"
+	"testing"
+	"time"
 )
 
 func TestRouter_ServeHTTP(t *testing.T) {
@@ -17,7 +19,44 @@ func TestRouter_ServeHTTP(t *testing.T) {
 
 	app := New()
 	server := app.HttpServer
-	r:=NewRouter(server)
+	r := NewRouter(server)
 
-	r.ServeHTTP(context.response.writer,context.request.Request)
+	r.ServeHTTP(context)
+}
+
+//
+func TestWrapRouterHandle(t *testing.T) {
+	param := &InitContextParam{
+		t,
+		"",
+		"",
+		test.ToDefault,
+	}
+
+	context := initAllContext(param)
+
+	app := New()
+	server := app.HttpServer
+	router := server.Router().(*router)
+	//use default config
+	server.SetSessionConfig(session.NewDefaultRuntimeConfig())
+	handle := router.wrapRouterHandle(Index, false)
+
+	handle(context)
+}
+
+func TestLogWebsocketContext(t *testing.T) {
+	param := &InitContextParam{
+		t,
+		"",
+		"",
+		test.ToDefault,
+	}
+
+	context := initAllContext(param)
+
+	log := logWebsocketContext(context, time.Now().Unix())
+	t.Log("logContext:", log)
+	//test.NotNil(t,log)
+	test.Equal(t, "", "")
 }
